+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Doc: fix missing `Timer` documentation following previous refactoring
 - Loader: fix loading/preloading of base64 audio assets, and base64 encoded FontFace
 - Text: fix an uncaught exception when removing Text object from the game world that use offScreenCanvas texture caching (thanks @wpernath)
+- Renderer: fix a regression with the masking features in the Canvas Rendering mode
 
 ## [12.0.0] (melonJS 2) - _2022-06-27_